Move stuff into middleware.py and permissions.py
(Moving code around only; it should have no functional changes)
